Chat rooms with bot or not

+++ src/main/java/org/lucifer/abchat/dao/ChatDao.java
@@ -5,5 +5,5 @@ import org.lucifer.abchat.domain.Chat;
 import java.util.List;
 
 public interface ChatDao extends BaseDao<Chat> {
-    List<Integer> freeRooms();
+    List<Long> freeRooms();
 }
